+++ include/elio/runtime/async_main.hpp
@@ -2,7 +2,6 @@
 
 #include "scheduler.hpp"
 #include <elio/coro/task.hpp>
-#include <elio/coro/vthread_stack.hpp>
 #include <atomic>
 #include <condition_variable>
 #include <csignal>
@@ -193,18 +192,11 @@ auto run(F&& f, const run_config& config = {})
     auto bound = [&f]() { return std::invoke(std::forward<F>(f)); };
 
     {
-        // Create root vthread_stack and set as current context
-        auto root_vstack = std::make_unique<coro::vthread_stack>();
-        auto* root_vstack_ptr = root_vstack.get();
-
-        auto wrapper = [&] {
-            // Construct coroutine (will allocate from root_vstack)
-            coro::vthread_stack_scope vstack_scope(root_vstack_ptr);
-            return detail::completion_wrapper<T>(std::move(bound), &signal);
-        }();
-        
+        auto* old_frame = coro::promise_base::current_frame();
+        auto wrapper = detail::completion_wrapper<T>(std::move(bound), &signal);
         auto handle = coro::detail::task_access::release(wrapper);
-        handle.promise().set_vstack_owner(root_vstack.release());
+        handle.promise().detach_from_parent();
+        coro::promise_base::set_current_frame(old_frame);
         sched.spawn(handle);
     }